+++ Changelog.md
@@ -10,6 +10,8 @@ Ultimately you _should_ notice less quirky behaviour at times which means you ca
 
 - Show the segment id when hovering over the segment name in the route list
 - The window state (size and position) are now stored in the user preferences
+- Show sprints, KOMs and Strava segments along the route on the main screen [#51](https://github.com/sandermvanvliet/RoadCaptain/issues/51)
+![Screenshot of Road Captain route builder showing special segments](./images/routebuilder-markers.png)
 
 ### Runner